+++ enrolment/helpers.py
@@ -2,21 +2,24 @@ import collections
 from http import cookies
 import re
 
-from directory_api_client.client import api_client
-from directory_ch_client.client import ch_search_api_client
-from directory_constants import choices
-from directory_sso_api_client.client import sso_api_client
+from directory_api_client import api_client
+from directory_ch_client import ch_search_api_client
+from directory_constants import choices, urls, user_roles
 from directory_forms_api_client import actions
-from directory_constants import urls
+from directory_sso_api_client import sso_api_client
 import directory_components
 
+from django.core.cache import cache
 from django.utils import formats
 from django.utils.dateparse import parse_datetime
 from django.conf import settings
 
+from enrolment import constants
+
+
 COMPANIES_HOUSE_DATE_FORMAT = '%Y-%m-%d'
-SESSION_KEY_COMPANY_PROFILE = 'COMPANY_PROFILE'
-SESSION_KEY_IS_ENROLLED = 'IS_ENROLLED'
+CACHE_KEY_COMPANY_PROFILE = 'COMPANY_PROFILE'
+CACHE_KEY_IS_ENROLLED = 'IS_ENROLLED'
 
 
 ProgressIndicatorConf = collections.namedtuple(
@@ -42,24 +45,15 @@ def claim_company(enrolment_key, personal_name, sso_session_id):
     response.raise_for_status()
 
 
-def get_company_profile(number, session):
-    session_key = f'{SESSION_KEY_COMPANY_PROFILE}-{number}'
-    if session_key not in session:
+def get_companies_house_profile(number):
+    key = f'{CACHE_KEY_COMPANY_PROFILE}-{number}'
+    value = cache.get(key)
+    if not value:
         response = ch_search_api_client.company.get_company_profile(number)
         response.raise_for_status()
-        session[session_key] = response.json()
-    return session[session_key]
-
-
-def create_user(email, password):
-    response = sso_api_client.user.create_user(email, password)
-    if response.status_code == 400:
-        # Check for non-password errors and ignore since we want to proceed
-        # For example we don't want to inform user of existing accounts
-        if not response.json().get('password'):
-            return None
-    response.raise_for_status()
-    return response.json()
+        value = response.json()
+        cache.set(key=key, value=value, timeout=60*60)
+    return value
 
 
 def user_has_company(sso_session_id):
@@ -71,16 +65,18 @@ def user_has_company(sso_session_id):
     response.raise_for_status()
 
 
-def get_is_enrolled(company_number, session):
-    session_key = f'{SESSION_KEY_IS_ENROLLED}-{company_number}'
-    if session_key not in session:
+def get_is_enrolled(company_number):
+    key = f'{CACHE_KEY_IS_ENROLLED}-{company_number}'
+    value = cache.get(key)
+    if not value:
         response = api_client.company.validate_company_number(company_number)
         if response.status_code == 400:
-            session[session_key] = True
+            value = True
         else:
             response.raise_for_status()
-            session[session_key] = False
-    return session[session_key]
+            value = False
+        cache.set(key=key, value=value, timeout=60*60)
+    return value
 
 
 def create_company_profile(data):
@@ -118,7 +114,7 @@ def notify_already_registered(email, form_url):
     response = action.save({
         'login_url': settings.SSO_PROXY_LOGIN_URL,
         'password_reset_url': settings.SSO_PROXY_PASSWORD_RESET_URL,
-        'contact_us_url': urls.FEEDBACK,
+        'contact_us_url': urls.domestic.FEEDBACK,
     })
 
     response.raise_for_status()
@@ -161,9 +157,34 @@ def collaborator_request_create(company_number, email, name, form_url):
         'name': name,
         'email': email,
         'collaborator_create_url': settings.FAB_ADD_USER_URL,
-        'report_abuse_url': urls.FEEDBACK,
+        'report_abuse_url': urls.domestic.FEEDBACK,
     })
     response.raise_for_status()
+
+
+def get_company_admins(sso_session_id):
+    response = api_client.company.collaborator_list(sso_session_id=sso_session_id)
+    response.raise_for_status()
+    collaborators = response.json()
+    return [collaborator for collaborator in collaborators if collaborator['role'] == user_roles.ADMIN]
+
+
+def create_company_member(data):
+    response = api_client.company.collaborator_create(data={**data, 'role': user_roles.MEMBER})
+    response.raise_for_status()
+
+
+def notify_company_admins_member_joined(sso_session_id, email_data, form_url):
+    company_admins = get_company_admins(sso_session_id)
+    assert company_admins, f"No admin found for {email_data['company_name']}"
+    for admin in company_admins:
+        action = actions.GovNotifyEmailAction(
+            email_address=admin['company_email'],
+            template_id=settings.GOV_NOTIFY_NEW_MEMBER_REGISTERED_TEMPLATE_ID,
+            form_url=form_url
+        )
+        response = action.save(email_data)
+        response.raise_for_status()
 
 
 class CompanyParser(directory_components.helpers.CompanyParser):
@@ -207,3 +228,18 @@ def parse_set_cookie_header(cookie_header):
     for cookie in split:
         simple_cookies.load(cookie)
     return simple_cookies
+
+
+def collaborator_invite_retrieve(invite_key):
+    response = api_client.company.collaborator_invite_retrieve(invite_key=invite_key)
+    if response.status_code == 200:
+        return response.json()
+
+
+def collaborator_invite_accept(sso_session_id, invite_key):
+    response = api_client.company.collaborator_invite_accept(sso_session_id=sso_session_id, invite_key=invite_key)
+    response.raise_for_status()
+
+
+def is_companies_house_details_incomplete(company_number):
+    return any(company_number.startswith(prefix) for prefix in constants.COMPANY_NUMBER_PREFIXES_INCOMPLETE_INFO)